P9_02_azure_function/requirements.txt
# Flask API requirements
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.0.0

# ML dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
joblib>=1.0.0

# Azure Functions (optional)
azure-functions>=1.11.0
